week1/python-full-code/requirements.txt
requests
beautifulsoup4
python-dotenv
openai
ipython
